requirements.txt
Flask>=2.3.0
gunicorn>=21.2.0
spotipy>=2.23.0
python-dotenv>=1.0.0